+++ src/cli/CommandHandler.h
@@ -29,15 +29,6 @@ SOFTWARE.
 
 class Print;
 
-/**
- * Increment argv and decrement argc so that it appears as if the command line
- * arguments are shifted to the *left* by one position. This is useful for
- * processing the command line tokens one at a time. Use
- * CommandHandler::isArgEqual() method to test for match against an expected
- * token.
- */
-#define SHIFT_ARGC_ARGV(argc, argv) do { argc--; argv++; } while (false)
-
 namespace ace_utils {
 namespace cli {
 
@@ -71,6 +62,23 @@ class CommandHandler {
     ace_common::FCString getHelpString() const { return mHelpString; }
 
   protected:
+    /**
+    * Increment argv and decrement argc so that it appears as if the command
+    * line arguments are shifted to the *left* by one position. This is useful
+    * for processing the command line tokens one at a time. Use
+    * CommandHandler::isArgEqual() method to test for match against an expected
+    * token.
+    */
+    static void shiftArgcArgv(int& argc, const char* const*& argv) {
+      argc--;
+      argv++;
+    }
+
+    /** Backwards compatibility with previous macro implementation . */
+    static void SHIFT_ARGC_ARGV(int& argc, const char* const*& argv) {
+      shiftArgcArgv(argc, argv);
+    }
+
     /** Test for equality against token. */
     static bool isArgEqual(const char* arg, const char* token) {
       return strcmp(arg, token) == 0;