+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
 * Unreleased
+* 0.4.1 (2021-01-22)
+    * Convert `SHIFT_ARGC_ARGV()` macro in `src/cli/CommandHandler.h` to
+      an inlined static function `CommandHandler::shiftArgcArgv()` using
+      references. I forgot that I was programming in C++ when I created the
+      macro. Add `SHIFT_ARGC_ARGV()` function for backwards compatibility.
+    * Update UnixHostDuino 0.4 to EpoxyDuino 0.5.
+    * No functional change in this release.
 * 0.4 (2020-12-09)
     * update `AceUtilsCrcEeprom.h`
         * Replace FastCRC with AceCRC library for a reduction of 4-5 kB of flash